notary signing support in SDK
Instead of relying on core we pull in the required signature support for notarization in SDK without any reliance on external processes.

1) getnotarysighash - gives sighash to sign per asset for a transaction. The asset is likely one that a notary cares about.
2) signhash - will sign compact signature provided with WIF private key and sighash

Some supporting functions created to be able to get assets in a transaction so users can filter them according to whatever rules (notarize ones that it cares about) and then pass that object into getNotarizationSigHash to fill notarysighash in the asset objects and then pass the same object in signing and adding to the transaction which now should be populated with a notarysig object which server can send back to client.