fix sysx burn to nevm when exact utxo amount used
when exact amount is used then outputs.length is 1 and assetChangeIndex interferes with optimizeOutputs later which removes the wrong output. Since outputs.length == 1 it means we selected the exact right amount of asset and no change existed and so assetChangeIndex should actually be removed in that case, it shouldn't even be in there in the first place but coinselect adds it because mapAssets includes as the first output thats burned and happen to use address: changeAddress which triggers the bug, we should correct for it in the code instead of force a change up to the user or consuming lib.